backend/requirements.txt
# Flask API
flask==2.3.3
flask-cors==3.0.10

# Environment variables
python-dotenv==1.0.1

# HTTP requests
requests==2.31.0

# PDF parsing
PyPDF2==3.0.1

# LangChain + Groq + LangGraph (lightweight compatible versions)
langchain==0.1.20
langchain-core==0.1.52
langchain-groq==0.1.3
langgraph==0.0.38

# For safe parsing of LLM structures
pydantic==1.10.13
